refactor: Default 유즈케이스 각각의 파일로 분리 및 그룹화
