+++ iOS/FlipMate/FlipMate/Domain/UseCases/Chart/DefaultFetchWeeklyChartUseCase.swift
@@ -1,0 +1,20 @@
+//
+//  DefaultFetchWeeklyChartUseCase.swift
+//  FlipMate
+//
+//  Created by 권승용 on 1/7/24.
+//
+
+import Foundation
+
+final class DefaultFetchWeeklyChartUseCase: FetchWeeklyChartUseCase {
+    private let repository: ChartRepository
+    
+    init(repository: ChartRepository) {
+        self.repository = repository
+    }
+    
+    func fetchWeeklyChartLog() async throws -> WeeklyChartLog {
+        return try await repository.fetchWeeklyLog()
+    }
+}
