+++ iOS/FlipMate/FlipMate/Domain/UseCases/DefaultValidateNicknameUseCase.swift
@@ -19,15 +19,3 @@ final class DefaultValidateNicknameUseCase: ValidateNicknameUseCase {
         return validationState
     }
 }
-
-final class DefaultSetupProfileInfoUseCase: SetupProfileInfoUseCase {
-    private let repository: ProfileSettingsRepository
-    
-    init(repository: ProfileSettingsRepository) {
-        self.repository = repository
-    }
-
-    func setupProfileInfo(nickName: String, profileImageData: Data) async throws -> UserInfo {
-        return try await repository.setupNewProfileInfo(nickName: nickName, profileImageData: profileImageData)
-    }
-}
