+++ iOS/FlipMate/FlipMate/Domain/UseCases/Timer/DefaultStartTimerUseCase.swift
@@ -19,18 +19,3 @@ final class DefaultStartTimerUseCase: StartTimerUseCase {
         return timerRepository.startTimer(startTime: startTime, categoryId: categoryId)
     }
 }
-
-final class DefaultFinishTimerUseCase: FinishTimerUseCase {
-    private let timerRepository: TimerRepsoitory
-    
-    init(timerRepository: TimerRepsoitory) {
-        self.timerRepository = timerRepository
-    }
-    
-    func finishTimer(endTime: Date, learningTime: Int, categoryId: Int?) -> AnyPublisher<Void, NetworkError> {
-        return timerRepository.finishTimer(
-            endTime: endTime,
-            learningTime: learningTime,
-            categoryId: categoryId)
-    }
-}
